+++ components/proxy/src/main/java/com/hotels/styx/startup/StyxServerComponents.java
@@ -36,7 +36,7 @@ import java.util.stream.StreamSupport;
 
 import static com.hotels.styx.Version.readVersionFrom;
 import static com.hotels.styx.infrastructure.logging.LOGBackConfigurer.initLogging;
-import static com.hotels.styx.startup.PluginLoadingForStartup.loadPlugins;
+import static com.hotels.styx.startup.extensions.PluginLoadingForStartup.loadPlugins;
 import static com.hotels.styx.startup.ServicesLoader.SERVICES_FROM_CONFIG;
 import static com.hotels.styx.startup.StyxServerComponents.LoggingSetUp.DO_NOT_MODIFY;
 import static java.util.Objects.requireNonNull;
