+++ components/proxy/src/main/java/com/hotels/styx/startup/extensions/DemoPlugin.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.startup;
+package com.hotels.styx.startup.extensions;
 
 import com.google.common.collect.ImmutableMap;
 import com.hotels.styx.api.Eventual;
